Fixed tinydtls bug in handshake failure
This patch is to fix bug of https://gerrit.iotivity.org/gerrit/#/c/7597/.

Issue : Double remove peer occur because removed peer in dtls_alert_send_from_err.

Patch 1,2: Call callback function in dtls_hadle_message when dtls handshake is failed.
         Destroy peer in case DLTS_CT_HANDSHAKE.
         Change checking peer point in handle_ccs
Patch 3,4: With patch 2, android-x86 build was failed. But, it is not related to this patch.
         So, adding just comment to re-start jenkins server.

Change-Id: Ie62ba1072dcbd2948d5931ed6fd606d540c650ba
Signed-off-by: js126.lee <js126.lee@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/7651
Reviewed-by: Chul Lee <chuls.lee@samsung.com>
Reviewed-by: Kyungsun Cho <goodsun.cho@samsung.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Randeep Singh <randeep.s@samsung.com>