Build: disable shared library of gtest as suggested.
As suggested by Jon A. Cruz "
 "GTest in general is more focused away from the shared.
  System-wide pre-compiled shared libs were deprecated a while back.
  Sets of programs built at the same time *might* be OK with a local shared version,
  but static linkage makes it safer."

So this patch explicitly disable dynamic library of gtest,
it can be reverted on trouble as there is no "hard requirement" for it.

For the record, I came to this once I noticed
that using default shared library
caused troubles on some Yocto distros (LfAgl) :

  relocation R_X86_64_PC32 \
  against undefined hidden symbol
  `__init_array_start'
  can not be used  ...

I suspect this problem is not related to gtest or iotivity.

Bug: https://jira.iotivity.org/browse/IOT-863
Bug-AGL: https://jira.automotivelinux.org/browse/SPEC-158
Change-Id: I9e4798aa3a579020255e8f53a8da8926dc4dee4f
Signed-off-by: Philippe Coval <philippe.coval@osg.samsung.com>