build: Add install feature for downstream
This helper makefile does not aim to replace scons,
but is an helper to rely on scons
which may not be supported out of the box
my packaging systems, this way many packaging tasks
can be factorized here between OS
to avoid duplication of efforts,
in hope they will be aligned as much as possible.

For now, this applies to Tizen , Yocto and Debian.

Bug: https://jira.iotivity.org/browse/IOT-524
Change-Id: Ice8f4de89ffdaad5a83aa80ca266edc836fbf6a0
Forewarded: https://gerrit.iotivity.org/gerrit/#/c/11813/
Origin: https://github.com/TizenTeam/iotivity/tree/sandbox/pcoval/on/master/fedora
Signed-off-by: Philippe Coval <philippe.coval@osg.samsung.com>