Fix for "UNUSED" redefined warning
The #define UNUSED is present in openjdk's jni.h and
in the arduino libraries. Resulting in the compilar
warning.

We already have OC_UNUSED defined in octypes.h which
was already avalible for all of the files that were
using UNUSED from logger.h.

The #define UNUSED was deleted from logger.h.
Any place UNUSED was used was renamed to OC_UNUSED.

Change-Id: If94ac13dc29eb3323e036eec750a18893895fb3d
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/13353
Reviewed-by: Larry Sachs <larry.j.sachs@intel.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Rick Bell <richard.s.bell@intel.com>