Fixed sign-compare warning
In the ParseChain function the bufLen and pos varaible were
changed from int to size_t.  Every location that ParseChain
is called is passing in a size_t as the bufLen an unsigned
type. The pos variable is never negative and should be the
same type as the bufLen.

While in the function the documentation was cleaned up and
comemnts were added to quickly identify the contents of the
pemCertHeader and pemCerFooter byte arrays. As well as clarify
the need of the goto `exit:` label when using the
CHECK_MBEDTLS_RET macro.

Change-Id: I1f063dddea29a2a00ae6e47c7b2240794bcaf5bb
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/13519
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Dmitriy Zhuravlev <d.zhuravlev@samsung.com>
Reviewed-by: Rick Bell <richard.s.bell@intel.com>