Remove not required OCStopMulticastServer() function call in RD
this is required but it is being moved to app layer, as it is
conflicting with the local and cloud scenarios. It would require
update to the local rd sample to call this function instead.
Also OCStopMulticastServer has to be made available to the
application developer, right now it is just a C API.

Change-Id: I7a51c29b7bc88416df81966f9d98898bc2b3ab08
Signed-off-by: hyuna0213.jo <hyuna0213.jo@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/13645
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jihun Ha <jihun.ha@samsung.com>
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Reviewed-by: Jaehong Jo <jaehong.jo@samsung.com>
Reviewed-by: Habib Virji <habib.virji@samsung.com>