Update the list of IPv6 multicast groups we join
According to OCF internal discussions, the upcoming OIC 1.1 spec will
have an errata that specifies that devices should always join the OCF
group on scopes 2, 3 and 5, and that devices connected to WiFi and/or
Ethernet should send discoveries on scope 2. IoTivity already sends on
scope 2 only; this patch merely enables joining on the other scopes.

Change-Id: Ic46ff326a6ba46bc877cfffd1483478d310e719b
Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/13963
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Dave Thaler <dthaler@microsoft.com>
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>