Addressed "missing braces around initializer" warning.
Gcc exhibited a "missing braces around initializer" warning when zero
initializing the ipv6mr_multiaddr field in the ipv6_mreq structure.
This appears to be a known issue with gcc.  Avoid the warning by
removing the explicit zero initialization, and rely on the implicit
initialization behavior of designated initializers instead.

Change-Id: I5658fb620433bc7f47a31d94bab2e9f07ce66f40
Signed-off-by: Ossama Othman <ossama.othman@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/14257
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Mike Fenelon <mike.fenelon@microsoft.com>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>