Use RTM_NEWADDR, RTM_DELADDR to receive info about an IP address.
We use RTM_NEWLINK multicast group mask to receive information about
a created or removed network interface. and If a specific network
interface is added, we call getifaddrs() to get an IP address
associated with an interface. but it is more correct to use
RTM_NEWADDR, RTM_DELADDR to detect IP address changes.

Change-Id: I9389cac95473583a015dd4d40489943dc856e0fd
Signed-off-by: hyuna0213.jo <hyuna0213.jo@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/14529
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: byonggon chun <bg.chun@samsung.com>
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Reviewed-by: Jaehong Jo <jaehong.jo@samsung.com>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>
Reviewed-by: Phil Coval <philippe.coval@osg.samsung.com>
Reviewed-by: Jaewook Jung <jw0213.jung@samsung.com>