[IOT-1570] fixed a crash issue related to block-wise transfer
When a resource server receives 4.08 error response from a resource
client, the server will try to send the response from the first
block again.
But now, the server assumes the message type that has to be sent
for receiving 4.08 error is only request message.
So I modified to handle it up to the message type.

Change-Id: I555b5dd5c44b73647c2543de551aba6f5113f909
Signed-off-by: Jaewook Jung <jw0213.jung@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/14433
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>