CTT1.2.2: changed requestHandle value to a request ID
When a request message is delivered to resource server, the address value
of serverRequest get passed as a requestHandle to the application.
However, when duplicated messages are delivered and the application is
doing slow-response on entityHandler, the later messages could have a
invalid address value, which is requestHandle, since the original
serverRequest would be deleted by responding to the first one.

So I made a unique ID for each request message and replace the requestHandle
with it.

Change-Id: I164bfeaea51be863715f7ced60c40b1f16dafdfc
Signed-off-by: Jaewook Jung <jw0213.jung@samsung.com>
Signed-off-by: jihwan.seo <jihwan.seo@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/14761
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Tested-by: Uze Choi <uzchoi@samsung.com>