Add ByteString type at the Value of RCSResourceAttributes.
ByteString(Binary) was added in the Value of OCRepresentation.
But, RE layer does not covered ByteString type.
In this patch, ByteString type is added in RE layer
for synchronization with OC.

patchset #2
handling of OC::AttributeType::Binary is added.
toString method of RCSByteString is fixed.

patchset #4
Description of RCSByteString is added.

patchset #5
Server sample application is update for including byte string
on GET response of custom interface.

Change-Id: If293954c64fe1b072b1bdff967681d6c6e774671
Signed-off-by: KIM JungYong <jyong2.kim@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/15695
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>