[4.0] Unification: Remove Profile Build Dependency & Prepare 4.0 Build
This is for Tizen 4.0.
 - Removed profile build dependency.
 - Prepare for Tizen 4.0 build.

a. Remove profile build dependency
 - This commit completely removes the build dependencies on Tizen profile.

b. Prepare for Tizen 4.0 build
 - We have assumed that Tizen 4.0 code is going to be more similar with 3.0
   than 2.0. Please do not rely on string '==' compare for version check.
   And, refer to the official macros that are defined in Project Config.
   https://build.tizen.org/project/prjconf?project=<project_name>

@To maintainer:
Please, keep a spec file syntax that was written at an official webpage.
https://docs.fedoraproject.org/ro/Fedora_Draft_Documentation/0.1/html/\
RPM_Guide/ch-specfile-syntax.html
First, if you want to use a comparison operator in %if, don't use
a double-quotes. Second, RPM macro does not support a floating-point
number for the comparison operator.

Change-Id: I8904a25e5e24f6e1bdbd460e434ad563695a3001
Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>